+++ sdk/trade360-java-sdk/src/main/java/eu/lsports/trade360_java_sdk/common/entities/markets/Market.java
@@ -1,5 +1,6 @@
 package eu.lsports.trade360_java_sdk.common.entities.markets;
 
+import com.fasterxml.jackson.annotation.JsonAlias;
 import eu.lsports.trade360_java_sdk.common.entities.enums.MarketStatus;
 import jakarta.annotation.Nullable;
 
@@ -40,5 +41,5 @@ public class Market {
     /**
      * Authoritative market status after TRADE calculation: 1=Open, 2=Suspended, 3=Settled.
      */
-    @Nullable public MarketStatus status;
+    @Nullable @JsonAlias("Status") public MarketStatus marketStatus;
 }